Fixing #729 - replacing yahoo with fixer

+++ admin/model/localisation/currency.php
@@ -126,8 +126,10 @@ class ModelLocalisationCurrency extends Model {
 
     public function refresh($force = false) {
         if (extension_loaded('curl')) {
-            $data = array();
-
+            $currencies = array();
+        
+            $default = $this->config->get('config_currency');
+        
             if ($force) {
                 $query = $this->db->query("SELECT * FROM " . DB_PREFIX . "currency WHERE code != '" . $this->db->escape($this->config->get('config_currency')) . "'");
             } else {
@@ -135,36 +137,46 @@ class ModelLocalisationCurrency extends Model {
             }
 
             foreach ($query->rows as $result) {
-                $data[] = $this->config->get('config_currency') . $result['code'] . '=X';
-            }
-
-            $curl = curl_init();
-
-            curl_setopt($curl, CURLOPT_URL, 'http://download.finance.yahoo.com/d/quotes.csv?s=' . implode(',', $data) . '&f=sl1&e=.csv');
-            curl_setopt($curl, CURLOPT_RETURNTRANSFER, 1);
-            curl_setopt($curl, CURLOPT_HEADER, false);
-            curl_setopt($curl, CURLOPT_CONNECTTIMEOUT, 30);
-            curl_setopt($curl, CURLOPT_TIMEOUT, 30);
-
-            $content = curl_exec($curl);
-
-            curl_close($curl);
-
-            $lines = explode("\n", trim($content));
-
-            foreach ($lines as $line) {
-                $currency = utf8_substr($line, 4, 3);
-                $value = utf8_substr($line, 11, 6);
-
-                if ((float)$value) {
-                    $this->db->query("UPDATE " . DB_PREFIX . "currency SET value = '" . (float)$value . "', date_modified = '" .  $this->db->escape(date('Y-m-d H:i:s')) . "' WHERE code = '" . $this->db->escape($currency) . "'");
+                if (($result['code'] != $default)) {
+                    $currencies[] = $result;
                 }
             }
 
-            $this->db->query("UPDATE " . DB_PREFIX . "currency SET value = '1.00000', date_modified = '" .  $this->db->escape(date('Y-m-d H:i:s')) . "' WHERE code = '" . $this->db->escape($this->config->get('config_currency')) . "'");
+            if($currencies){
+                $curl = curl_init();
 
-            $this->cache->delete('currency');
+                curl_setopt($curl, CURLOPT_URL, 'https://api.fixer.io/latest?base=' . $default);
+                curl_setopt($curl, CURLOPT_RETURNTRANSFER, 1);
+                curl_setopt($curl, CURLOPT_HEADER, false);
+                curl_setopt($curl, CURLOPT_SSL_VERIFYPEER, 0);
+                curl_setopt($curl, CURLOPT_CONNECTTIMEOUT, 30);
+                curl_setopt($curl, CURLOPT_TIMEOUT, 30);
+
+                $response = curl_exec($curl);
+
+                curl_close($curl);
+
+                $response_info = json_decode($response, true);
+
+                if (isset($response_info['rates'])) {
+                    foreach ($currencies as $currency) {
+                        if (isset($response_info['rates'][$currency['code']])) {
+                            $this->editValueByCode($currency['code'], $response_info['rates'][$currency['code']]);
+                        }
+                    }
+                }
+
+                $this->cache->delete('currency');
+            }
+
+            $this->editValueByCode($default, '1.00000');
         }
+    }
+
+    public function editValueByCode($code, $value) {
+        $this->db->query("UPDATE " . DB_PREFIX . "currency SET value = '" . (float)$value . "', date_modified = NOW() WHERE code = '" . $this->db->escape((string)$code) . "'");
+
+        $this->cache->delete('currency');
     }
 
     public function getTotalCurrencies() {
